# docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Fairspace APIs
  description: Fairspace provides APIs for managing files and its associated metadata. Files can be managed using the WebDAV protocol at the path /webdav/v1. Each webdav resource has a special `iri` property indicating the iri that is used to describe its metadata.
  license:
    name: TBD
servers:
  - url: https://fairspace.ci.fairway.app/api/v1
paths:
...
  /projects/{project}/collections/:
    get:
      summary: Retrieve a list of collections
      operationId: getCollections
      tags:
        - collections
      parameters:
        - $ref: '#/components/parameters/projectParam'
        - name: iri
          in: query
          description: If given, a single collection with the given IRI will be returned
          required: false
          schema:
            type: string
            format: iri
      responses:
        '200':
          description: Succesful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ReturnedCollection'
                  - $ref: '#/components/schemas/Collections'
              examples:
                single:
                  value: |
                    {
                      "iri": "http://fairspace.ci.fairway.app/iri/3a95581c-de86-4977-a853-0237b3ee728e",
                      "dateCreated": "2019-04-24T13:49:05.547Z",
                      "createdBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                      "dateModified": "2019-04-24T13:49:05.547Z",
                      "modifiedBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                      "dateDeleted": null,
                      "deletedBy": null,
                      "name": "Collection update",
                      "description": "Description for test 人物",
                      "location": "aa",
                      "type": "LOCAL",
                      "access": "Manage",
                      "canWrite": true,
                      "canManage": true,
                      "canRead": true
                    }
                multiple:
                  value: |
                    [
                      {
                        "iri": "http://fairspace.ci.fairway.app/iri/3a95581c-de86-4977-a853-0237b3ee728e",
                        "dateCreated": "2019-04-24T13:49:05.547Z",
                        "createdBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                        "dateModified": "2019-04-24T13:49:05.547Z",
                        "modifiedBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                        "dateDeleted": null,
                        "deletedBy": null,
                        "name": "Collection update",
                        "description": "Description for test 人物",
                        "location": "aa",
                        "type": "LOCAL",
                        "access": "Manage",
                        "canWrite": true,
                        "canManage": true,
                        "canRead": true
                      },
                      {
                          "iri": "http://fairspace.ci.fairway.app/iri/f1487ce4-e4c4-4243-af34-18e33d0be7e5",
                          "dateCreated": "2019-04-24T14:14:18.093Z",
                          "createdBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                          "dateModified": "2019-04-24T14:14:18.093Z",
                          "modifiedBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                          "dateDeleted": null,
                          "deletedBy": null,
                          "name": "test",
                          "description": "",
                          "location": "test",
                          "type": "LOCAL_FILE",
                          "access": "Manage",
                          "canWrite": true,
                          "canManage": true,
                          "canRead": true
                      }
                    ]
        '404':
          description: Collection not found
    put:
      summary: Create new collection
      operationId: createCollection
      tags:
        - collections
      parameters:
        - $ref: '#/components/parameters/projectParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreation'
      responses:
        '201':
          description: Created collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnedCollection'
              example: |
                {
                  "iri": "http://fairspace.ci.fairway.app/iri/3a95581c-de86-4977-a853-0237b3ee728e",
                  "dateCreated": "2019-04-24T13:49:05.547Z",
                  "createdBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                  "dateModified": "2019-04-24T13:49:05.547Z",
                  "modifiedBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                  "dateDeleted": null,
                  "deletedBy": null,
                  "name": "Collection update",
                  "description": "Description for test 人物",
                  "location": "aa",
                  "type": "LOCAL",
                  "access": "Manage",
                  "canWrite": true,
                  "canManage": true,
                  "canRead": true
                }
        '409':
          description: Conflict - location is already in use
    patch:
      summary: Update collection
      operationId: updateCollection
      tags:
        - collections
      parameters:
        - $ref: '#/components/parameters/projectParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '201':
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnedCollection'
              example: |
                {
                  "iri": "http://fairspace.ci.fairway.app/iri/3a95581c-de86-4977-a853-0237b3ee728e",
                  "dateCreated": "2019-04-24T13:49:05.547Z",
                  "createdBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                  "dateModified": "2019-04-24T13:49:05.547Z",
                  "modifiedBy": "http://fairspace.ci.fairway.app/iri/6e6cde34-45bc-42d8-8cdb-b6e9faf890d3",
                  "dateDeleted": null,
                  "deletedBy": null,
                  "name": "Collection update",
                  "description": "Description for test 人物",
                  "location": "aa",
                  "type": "LOCAL",
                  "access": "Manage",
                  "canWrite": true,
                  "canManage": true,
                  "canRead": true
                }
        '404':
          description: Collection with the given IRI could not be found
        '409':
          description: Conflict - location is already in use
...
components:
  securitySchemes:
    OpenID:
      type: oauth2
      flows:
        password:
          tokenUrl: https://keycloak.hyperspace.ci.fairway.app/auth/realms/ci/protocol/openid-connect/token
  schemas:
...
    BasicCollection:
      properties:
        name:
          type: string
        description:
          type: string
        location:
          description: Location where the collection can be found. Represents the top-level directory
          type: string
        type:
          type: string
          enum:
            - LOCAL
    CollectionCreation:
      allOf:
        - $ref: '#/components/schemas/BasicCollection'
        - required:
            - name
            - location
            - type
    CollectionUpdate:
      allOf:
        - $ref: '#/components/schemas/BasicCollection'
        - properties:
            iri:
              type: string
              format: iri
        - required:
            - iri
    ReturnedCollection:
      allOf:
        - $ref: '#/components/schemas/BasicCollection'
        - $ref: '#/components/schemas/DerivedAccessProperties'
        - properties:
            iri:
              type: string
              format: iri
            name:
              type: string
            description:
              type: string
            location:
              description: Location where the collection can be found. Represents the top-level directory
              type: string
            type:
              type: string
              enum:
                - LOCAL
            createdBy:
              type: string
              format: iri
              readOnly: true
            dateCreated:
              type: string
              format: date-time
              readOnly: true
            modifiedBy:
              type: string
              format: iri
              readOnly: true
            dateModified:
              type: string
              format: date-time
              readOnly: true
            deletedBy:
              type: string
              format: iri
              readOnly: true
            dateDeleted:
              type: string
              format: date-time
              readOnly: true
            access:
              type: string
              enum: [Manage,Write,Read,None]
              readOnly: true
            canManage:
              type: boolean
              readOnly: true
            canRead:
              type: boolean
              readOnly: true
            canWrite:
              type: boolean
              readOnly: true
    Collections:
      type: array
      items:
        $ref: "#/components/schemas/ReturnedCollection"
    DerivedAccessProperties:
      properties:
        canManage:
          type: boolean
          readOnly: true
        canRead:
          type: boolean
          readOnly: true
        canWrite:
          type: boolean
          readOnly: true
...
  parameters:
    projectParam:
      name: project
      in: path
      description: The project
      required: true
      schema:
        type: string
...
security:
  - OpenID:
      - openid